+++ lib/shopify_cli.rb
@@ -116,6 +116,7 @@ module ShopifyCli
   autoload :Project, 'shopify-cli/project'
   autoload :ProjectType, 'shopify-cli/project_type'
   autoload :Resources, 'shopify-cli/resources'
+  autoload :Shopifolk, 'shopify-cli/shopifolk'
   autoload :SubCommand, 'shopify-cli/sub_command'
   autoload :Task, 'shopify-cli/task'
   autoload :Tasks, 'shopify-cli/tasks'