+++ test/project_types/extension/extension_test_helpers.rb
@@ -7,6 +7,7 @@ module Extension
     autoload :TestExtensionSetup, 'project_types/extension/extension_test_helpers/test_extension_setup'
     autoload :TempProjectSetup, 'project_types/extension/extension_test_helpers/temp_project_setup'
     autoload :Messages, 'project_types/extension/extension_test_helpers/messages'
+    autoload :DummyArgo, 'project_types/extension/extension_test_helpers/dummy_argo'
 
     module Stubs
       autoload :GetOrganizations, 'project_types/extension/extension_test_helpers/stubs/get_organizations'