+++ test/project_types/extension/extension_test_helpers/test_extension.rb
@@ -5,6 +5,10 @@ module Extension
     class TestExtension < Models::Type
       IDENTIFIER = 'TEST_EXTENSION'
 
+      def graphql_identifier
+        'TEST_EXTENSION_GQL'
+      end
+
       def name
         'Test Extension'
       end