+++ test/shopify-cli/commands/config_test.rb
@@ -10,6 +10,16 @@ module ShopifyCli
         run_cmd('config help')
       end
 
+      def test_feature_help_argument_calls_help
+        @context.expects(:puts).with(ShopifyCli::Commands::Config::Feature.help)
+        run_cmd('config feature --help')
+      end
+
+      def test_analytics_help_argument_calls_help
+        @context.expects(:puts).with(ShopifyCli::Commands::Config::Analytics.help)
+        run_cmd('config analytics --help')
+      end
+
       def test_no_arguments_calls_help
         @context.expects(:puts).with(ShopifyCli::Commands::Config.help)
         run_cmd('config')
@@ -20,16 +30,30 @@ module ShopifyCli
         run_cmd('config feature')
       end
 
-      def test_will_disable_a_feature_that_is_enabled
+      def test_will_enable_a_feature_that_is_disabled
         ShopifyCli::Feature.disable(TEST_FEATURE)
+        @context.expects(:puts).with(@context.message('core.config.feature.enabled', TEST_FEATURE))
         run_cmd("config feature #{TEST_FEATURE} --enable")
         assert ShopifyCli::Feature.enabled?(TEST_FEATURE)
       end
 
-      def test_will_enable_a_feature_that_is_disabled
+      def test_will_disable_a_feature_that_is_enabled
         ShopifyCli::Feature.enable(TEST_FEATURE)
+        @context.expects(:puts).with(@context.message('core.config.feature.disabled', TEST_FEATURE))
         run_cmd("config feature #{TEST_FEATURE} --disable")
         refute ShopifyCli::Feature.enabled?(TEST_FEATURE)
+      end
+
+      def test_will_enable_analytics_that_is_disabled
+        ShopifyCli::Config.set('analytics', 'enabled', false)
+        run_cmd("config analytics --enable")
+        assert ShopifyCli::Config.get_bool('analytics', 'enabled')
+      end
+
+      def test_will_disable_analytics_that_is_enabled
+        ShopifyCli::Config.set('analytics', 'enabled', true)
+        run_cmd("config analytics --disable")
+        refute ShopifyCli::Config.get_bool('analytics', 'enabled')
       end
     end
   end